feat: add spec step details drawer
