+++ ui/src/features/dags/components/step-details/index.ts
@@ -1,0 +1,1 @@
+export { StepDetails } from './StepDetails';
